refactor dataset and template hooks
