+++ plugin/web/extensions/sidebar/core/components/hooks/projectHooks.ts
@@ -3,10 +3,10 @@ import { Project, ReearthApi } from "@web/extensions/sidebar/types";
 import { generateID, mergeProperty, postMsg } from "@web/extensions/sidebar/utils";
 import { useCallback, useState } from "react";
 
-import { Data, DataCatalogItem, Template } from "../../types";
+import { DataCatalogItem, Template } from "../../types";
 import { FieldComponent } from "../content/common/DatasetCard/Field/Fields/types";
 
-import { mergeOverrides } from ".";
+import { mergeOverrides } from "./utils";
 
 export const defaultProject: Project = {
   sceneOverrides: {
@@ -44,27 +44,7 @@ export const defaultProject: Project = {
   userStory: undefined,
 };
 
-export default ({
-  data,
-  fieldTemplates,
-  backendURL,
-  backendProjectName,
-  backendAccessToken,
-  inEditor,
-  processedCatalog,
-  setLoading,
-  handleBackendFetch,
-}: {
-  data?: Data[];
-  fieldTemplates?: Template[];
-  backendURL?: string;
-  backendProjectName?: string;
-  backendAccessToken?: string;
-  inEditor?: boolean;
-  processedCatalog: DataCatalogItem[];
-  setLoading?: React.Dispatch<React.SetStateAction<boolean>>;
-  handleBackendFetch: () => Promise<void>;
-}) => {
+export default ({ fieldTemplates }: { fieldTemplates?: Template[] }) => {
   const [project, updateProject] = useState<Project>(defaultProject);
   const [cleanseOverride, setCleanseOverride] = useState<string>();
 
@@ -215,107 +195,6 @@ export default ({
     postMsg({ action: "removeAllDatasetsFromScene" });
   }, []);
 
-  const handleDatasetUpdate = useCallback(
-    (updatedDataset: DataCatalogItem, cleanseOverride?: any) => {
-      updateProject(project => {
-        const updatedDatasets = [...project.datasets];
-        const datasetIndex = updatedDatasets.findIndex(d2 => d2.dataID === updatedDataset.dataID);
-        if (datasetIndex >= 0) {
-          if (updatedDatasets[datasetIndex].visible !== updatedDataset.visible) {
-            postMsg({
-              action: "updateDatasetVisibility",
-              payload: { dataID: updatedDataset.dataID, hide: !updatedDataset.visible },
-            });
-          }
-          if (cleanseOverride) {
-            setCleanseOverride(cleanseOverride);
-          }
-          updatedDatasets[datasetIndex] = updatedDataset;
-        }
-        const updatedProject = {
-          ...project,
-          datasets: updatedDatasets,
-        };
-        postMsg({ action: "updateProject", payload: updatedProject });
-        return updatedProject;
-      });
-    },
-    [],
-  );
-
-  const handleDataRequest = useCallback(
-    async (dataset?: DataCatalogItem) => {
-      if (!backendURL || !backendAccessToken || !dataset) return;
-      const datasetToSave = convertToData(dataset);
-
-      const isNew = !data?.find(d => d.dataID === dataset.dataID);
-
-      const fetchURL = !isNew
-        ? `${backendURL}/sidebar/${backendProjectName}/data/${dataset.id}` // should be id and not dataID because id here is the CMS item's id
-        : `${backendURL}/sidebar/${backendProjectName}/data`;
-
-      const method = !isNew ? "PATCH" : "POST";
-
-      const res = await fetch(fetchURL, {
-        headers: {
-          authorization: `Bearer ${backendAccessToken}`,
-        },
-        method,
-        body: JSON.stringify(datasetToSave),
-      });
-      if (res.status !== 200) {
-        handleBackendFetch();
-        return;
-      }
-      const data2 = await res.json();
-      console.log("DATA JUST SAVED: ", data2);
-      handleBackendFetch(); // MAYBE UPDATE THIS LATER TO JUST UPDATE THE LOCAL VALUE
-    },
-    [data, backendAccessToken, backendURL, backendProjectName, handleBackendFetch],
-  );
-
-  const handleDatasetSave = useCallback(
-    (dataID: string) => {
-      (async () => {
-        if (!inEditor) return;
-        setLoading?.(true);
-        const selectedDataset = project.datasets.find(d => d.dataID === dataID);
-
-        await handleDataRequest(selectedDataset);
-        setLoading?.(false);
-      })();
-    },
-    [inEditor, project.datasets, setLoading, handleDataRequest],
-  );
-
-  const handleDatasetPublish = useCallback(
-    (dataID: string, publish: boolean) => {
-      (async () => {
-        if (!inEditor || !processedCatalog) return;
-        const dataset = processedCatalog.find(item => item.dataID === dataID);
-
-        if (!dataset) return;
-
-        dataset.public = publish;
-
-        updateProject(project => {
-          const updatedDatasets = [...project.datasets];
-          const datasetIndex = updatedDatasets.findIndex(d2 => d2.dataID === dataID);
-          if (datasetIndex >= 0) {
-            updatedDatasets[datasetIndex] = dataset;
-          }
-          return {
-            ...project,
-            datasets: updatedDatasets,
-          };
-        });
-
-        await handleDataRequest(dataset);
-      })();
-    },
-    [processedCatalog, inEditor, handleDataRequest],
-  );
-
   const handleOverride = useCallback(
     (dataID: string, activeIDs?: string[]) => {
       const dataset = project.datasets.find(d => d.dataID === dataID);
@@ -334,24 +213,11 @@ export default ({
   return {
     project,
     updateProject,
+    setCleanseOverride,
     handleProjectSceneUpdate,
     handleProjectDatasetAdd,
     handleProjectDatasetRemove,
     handleProjectDatasetRemoveAll,
-    handleDatasetUpdate,
-    handleDatasetSave,
-    handleDatasetPublish,
     handleOverride,
-  };
-};
-
-const convertToData = (item: DataCatalogItem): Data => {
-  return {
-    dataID: item.dataID,
-    public: item.public,
-    visible: item.visible ?? true,
-    template: item.template,
-    components: item.components,
-    fieldGroups: item.fieldGroups,
   };
 };
